requirements.txt
alembic==1.14.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.8.1
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
cryptography==43.0.3
cssselect2==0.7.0
Django==4.2
django-celery-beat==2.7.0
django-celery-results==2.5.1
django-classy-tags==3.0.1
django-formtools==2.4
django-js-asset==2.0.0
django-mptt==0.16.0
django-polymorphic==3.1.0
django-sekizai==3.0.1
django-timezone-field==7.0
django-treebeard==4.4
djangocms-admin-style==3.3.1
dnspython==2.7.0
easy-thumbnails==2.9
email_validator==2.2.0
fastapi==0.115.4
fastapi-cli==0.0.5
greenlet==3.1.1
h11==0.14.0
html5lib==1.1
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
idna==3.10
Jinja2==3.1.4
kombu==5.4.2
lxml==5.3.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector-python==9.1.0
mysqlclient==2.2.6
numpy==2.1.3
oracledb==2.5.0
packaging==24.2
pandas==2.2.3
pika==1.3.2
pillow==11.0.0
prompt_toolkit==3.0.48
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
Pygments==2.18.0
PyMySQL==1.1.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
PyYAML==6.0.2
reportlab==4.2.5
requests==2.32.3
rich==13.9.4
setuptools==75.6.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlparse==0.5.2
starlette==0.41.2
svglib==1.5.1
tinycss2==1.4.0
tqdm==4.67.0
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
urllib3==2.2.3
uvicorn==0.32.0
vine==5.1.0
watchfiles==0.24.0
wcwidth==0.2.13
webencodings==0.5.1
websockets==14.1
